Restricted `Assign()` Safety to Uninitialized Values
This will only avoid errors when assigning an uninitialized value: a scalar that is Empty or an object that is Nothing.  This works smoothly in all but two cases, where the `val` is scalar unwrapped by a Variant:
(1) when the `var` is an object wrapped by a Variant; or
(2) when the `var` is an unwrapped object.

Thus, it suffices for assigning to unwrapped `Dix_*()` fields the value returned by `Clx_Get()`, which is wrapped in a Variant.