Coordinated `AsObj()` with Class Initialization
Initially this was a bug that would surely (?) throw a syntax error for a missing `cls` argument.

Also updated logic flow to conceptually distinguish between a nonexistent class and a blank one.  Currently `Obj_Class()` returns a blank ("") in both cases, but now the updated logic would properly distinguish them, if we used (say) a NULL class to explicitly designate a simulated object as unspecified.