Privatized `*_Initialize()` Procedure in Template
Since `*_Initialize()` merely ensures the SOb (and its fields) are ready for manipulation by fellow procedures, it is by nature an internal sub.  So by default the template should not expose it to the end user, on behalf of our dev who is using the template.  If the dev wants to publicize their initialization sub, they may simply replace the Private keyword with Public.

* Also clarified comments so the phrase "enumeration" is solely linked to the Enum statement.