Realigned Table Checkmarks in README
Now checkmarks are separated from footnotes by a single space (" ").  And "orphan footnotes" (without a checkmark) are indented in their cell with an em-space (" "), separated from the pipe ("|") by two regular spaces (" "), which yields the following text: "|   [^".

The result is something that aligns all footnotes in both the markdown source and also the rendered webpage.